created admin auth or not middleware and implemented

+++ src/app/api/admin/delete-product/route.js
@@ -7,22 +7,32 @@ export const dynamic = "force-dynamic";
 export async function DELETE(req) {
   try {
     await connectToDB();
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
+    
+    const isAuthUser = await AuthUser(req);
 
-    if (!id)
-      return NextResponse.json({
-        success: false,
-        message: "Product ID is required",
-      });
+    if (isAuthUser?.role === "admin") {
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get("id");
 
-    const deletedProduct = await Product.findByIdAndDelete(id);
+      if (!id)
+        return NextResponse.json({
+          success: false,
+          message: "Product ID is required",
+        });
 
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: "Product deleted successfully",
-      });
+      const deletedProduct = await Product.findByIdAndDelete(id);
+
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: "Product deleted successfully",
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: "Failed to delete the product ! Please try again",
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
@@ -33,7 +43,7 @@ export async function DELETE(req) {
     console.log(e);
     return NextResponse.json({
       success: false,
-      message: "Something went wrong ! Please try again later",
+      message: "You are not Authenticated",
     });
   }
 }
